generate account in a separate script.
