+++ docs/sdk/1-workshop/03-attester/01-account.md
@@ -5,8 +5,7 @@ title: Account
 
 import CodeBlock from '@theme/CodeBlock';
 import SnippetBlock from '../../../../src/components/SnippetBlock';
-import Index from '!!raw-loader!../../../../code-examples/workshop/attester/index.js';
-import GetAccount from '!!raw-loader!../../../../code-examples/workshop/attester/getAccount.js';
+import GenerateAccount from '!!raw-loader!../../../../code-examples/workshop/attester/generateAccount.js';
 
 After you have [setup the project structure](./) in the last step, we'll create our <span class="label-role attester">Attester</span> account.
 In KILT, there is an account which is an object that interacts with the blockchain.
@@ -47,32 +46,14 @@ You'll be prompted to save it after the first run.
 Open `attester/getAccount.js` and paste the following code:
 
 <CodeBlock className="language-js" title="attester/getAccount.js">
-  {GetAccount}
+  {GenerateAccount}
 </CodeBlock>
 
 ## Execute
 
-Let's setup our <span class="label-role attester">Attester</span> index. Copy the below into `attester/index.js`
-
-<SnippetBlock
-  title="attester/index.js"
-  className="language-js"
-  snippets='[
-    [0, 11],
-    [12, 14],
-    [14,16],
-    [20,25],
-    "  console.log(account);",
-    [28,30],
-    "initialize();"
-  ]'
->
-  {Index}
-</SnippetBlock>
-
 Now run it to get your <span class="label-role attester">Attester</span> `<address>` and `<mnenomic>`.
 ```bash
-node ./attester/index.js
+node ./attester/generateAccount.js
 ```
 
 Your output will provide you with `ATTESTER_MNEMONIC` and `ATTESTER_ADDRESS`. Be sure to save it in your `.env`
